# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "StoreFy API",
        "description": "StoreFY Swagger OpenApi description",
        "contact": {
            "email": "georgeton.ufg@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Api local server"
        },
        {
            "url": "http://storefy.workingtech.com.br:31080/api",
            "description": "Api dev server"
        }
    ],
    "paths": {
...
        "/tipos-categoria": {
            "get": {
                "tags": [
                    "tipos-categoria"
                ],
                "summary": "Get all items",
                "operationId": "getTiposCategoria",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/orderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/TiposCategoria"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "tipos-categoria"
                ],
                "summary": "Add a new item",
                "operationId": "addTiposCategoria",
                "requestBody": {
                    "description": "Object to add new item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TiposCategoria"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/TiposCategoria"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "$ref": "#/components/schemas/validate_message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tipos-categoria/{id}": {
            "get": {
                "tags": [
                    "tipos-categoria"
                ],
                "summary": "Get item by ID",
                "operationId": "getTiposCategoria_id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "$ref": "#/components/parameters/fields"
                    },
                    {
                        "$ref": "#/components/parameters/where"
                    },
                    {
                        "$ref": "#/components/parameters/orderBy"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/TiposCategoria"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "tipos-categoria"
                ],
                "summary": "Update item",
                "operationId": "updateTiposCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Object to update item",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TiposCategoria"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "allOf": [
                                        {
                                            "properties": {
                                                "id": {
                                                    "title": "id",
                                                    "description": "id",
                                                    "type": "integer"
                                                }
                                            }
                                        },
                                        {
                                            "$ref": "#/components/schemas/TiposCategoria"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "tipos-categoria"
                ],
                "summary": "Delete item",
                "operationId": "deleteTiposCategoria",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "validate_message": {
                "title": "Error message",
                "description": "Error jso1n",
                "required": [
                    "code",
                    "message"
                ],
                "properties": {
                    "code": {
                        "type": "string",
                        "readOnly": true
                    },
                    "message": {
                        "type": "string",
                        "minLength": 8,
                        "readOnly": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorMessage"
                }
            },
...
            "TiposCategoria": {
                "title": "TiposCategoria model",
                "description": "TiposCategoria model",
                "required": [],
                "properties": {
                    "descricao": {
                        "title": "descricao",
                        "description": "Descrição"
                    },
                    "codigo": {
                        "title": "codigo",
                        "description": "Codigo"
                    },
                    "endPoint": {
                        "title": "endPoint",
                        "description": "Endpoint"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TiposCategoria"
                }
            }
        },
        "parameters": {
            "fields": {
                "name": "fields",
                "in": "query",
                "description": "Fields to return"
            },
            "where": {
                "name": "where",
                "in": "query",
                "description": "filter of any database fields (always send ternaries: 'field,operator,value')'"
            },
            "orderBy": {
                "name": "orderBy",
                "in": "query",
                "description": "Order ny database fields (always send: 'field,order')'"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}